Add update in storage.js file in module
